ATR-767: moved refactoring and its tests to the analyzers subfolder and namespace
