+++ src/Acuminator/Acuminator.Tests/Tests/StaticAnalysis/NameConventionEventsInGraphsAndGraphExtensions/Sources/CacheAttached.cs
