+++ src/Acuminator/Acuminator.Tests/Tests/StaticAnalysis/NameConventionEventsInGraphsAndGraphExtensions/Sources/CacheAttached_Expected.cs
