+++ src/Acuminator/Acuminator.Tests/Tests/StaticAnalysis/NameConventionEventsInGraphsAndGraphExtensions/Sources/FieldEventHandler.cs
