+++ src/Acuminator/Acuminator.Tests/Tests/StaticAnalysis/NameConventionEventsInGraphsAndGraphExtensions/Sources/Override.cs
